fix: problem when delete database
