--- system/fs/header.mjs
@@ -1,3 +1,0 @@
-(() => {
-    
-})();
